refactor: PochiPredictor の継承を解消
