+++ pochitrain/inference/__init__.py
@@ -1,8 +1,8 @@
 """
 Pochitrainの推論サポートモジュール.
 
-推論に関連するCSV出力機能を提供します。
-メインの推論機能は pochitrain.pochi_predictor を参照してください。
+推論に関連するCSV出力機能を提供します.
+メインの推論機能は pochitrain.pochi_predictor を参照してください.
 """
 
 from .csv_exporter import InferenceCSVExporter
